Mamorx/services/mamorx_service/pyproject.toml
[tool.poetry]
name = "mamorx-service"
version = "0.1.0"
description = ""
authors = ["Pawin Taechoyotin <pawin.taechoyotin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"
crewai = {version = "0.51.1", extras = ["tools"]}
anthropic = "^0.37.1"
langchain-aws = "0.1.17"
grobid-client-python = "^0.0.8"
lxml = "^5.3.0"
bs4 = "^0.0.2"
ratelimit = "^2.2.1"
boto3 = "1.34.162"
fastapi = {extras = ["standard"], version = "^0.115.4"}
pydantic-settings = "^2.6.0"
redis = "^5.2.0"
python-multipart = "^0.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
